requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo==4.6.1
scikit-learn
dill
pymal

##-e .